Add feature line shader
The shader traces secondary rays in a disc around the current ray and
passes them through a series of edge tests, to get a count of the
number rays near an edge. This count is then used to blend the objects
color with a feature line color.

The edges supported are:

* Intersections marking curves between two or more primitives
* Silhoutte edges when a primitives normal is perpendicular to the
  viewing direction
* Self silhoutte edges where there is a discontinuity in the objects
  normals field.

Based on "Ray Tracing NPR-Style Feature Lines", differing on the test
for a silhoutte edge.